make context from UA/hints functions associated
