+++ relay-general/src/protocol/contexts/browser.rs
@@ -1,4 +1,6 @@
+use crate::store::user_agent::{get_version, is_known};
 use crate::types::{Annotated, Object, Value};
+use crate::user_agent::{parse_user_agent, RawUserAgentInfo};
 
 /// Web browser information.
 #[derive(Clone, Debug, Default, PartialEq, Empty, FromValue, IntoValue, ProcessValue)]
@@ -19,5 +21,28 @@ impl BrowserContext {
     /// The key under which a browser context is generally stored (in `Contexts`)
     pub fn default_key() -> &'static str {
         "browser"
+    }
+
+    pub fn new_from_client_hints(raw_contexts: &RawUserAgentInfo) -> Option<BrowserContext> {
+        let browser = raw_contexts.sec_ch_ua?.to_owned();
+
+        Some(BrowserContext {
+            name: Annotated::new(browser),
+            ..Default::default()
+        })
+    }
+
+    pub fn new_from_user_agent(user_agent: &str) -> Option<BrowserContext> {
+        let browser = parse_user_agent(user_agent);
+
+        if !is_known(browser.family.as_str()) {
+            return None;
+        }
+
+        Some(BrowserContext {
+            name: Annotated::from(browser.family),
+            version: Annotated::from(get_version(&browser.major, &browser.minor, &browser.patch)),
+            ..BrowserContext::default()
+        })
     }
 }
